% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{AbCdEf.BM}
\alias{AbCdEf.BM}
\title{Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree}
\usage{
\method{AbCdEf}{BM}(tree, model, metaI = validateModel.BM(tree, model, verbose
  = verbose), pc, verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object (see details)}

\item{model}{parameters of the BM process. This must be a
named list with the following elements:
Sigma: a R x k x k array, each Sigma[r,,] containing the
matrix Sigma for regime r;
Sigmae: a R x k x k array, each Sigmae[r,,] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{presentCoords}{a M x k logical matrix representing the present coordinates at each
node}
}
\value{
a named list containing the following elements:
A: a M x k x k array, A[i,,] corresponding to Ai for
each branch ending at node or tip i;
b: a M x k matrix, b[i,] corresponding to the vectors bi;
C: a M x k x k array, C[i,,] corresponding to the
matrices Ci;
d: a M x k matrix, d[i,] corresponding to the vectors di;
E: a M x k x k array, E[i,,] corresponding to the matrices Ei;
f: a vector, f[i] correspondign to fi
}
\description{
Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree
}
\details{
The dimnames
}
