% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{AbCdEf.JOU}
\alias{AbCdEf.JOU}
\title{Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree}
\usage{
\method{AbCdEf}{JOU}(tree, model, metaI = validateModel.JOU(tree, model,
  verbose = verbose), pc, verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object (see details)}

\item{model}{parameters of the JOU process. This must be a
named list with the following elements:
H: a k x k x R array, where R is the number of regimes of the
JOU process, k is the number of variables (traits), each H[,,r]
containing the matrix H for regime r;
Theta: a k x R matrix, row Theta[, r] containing the long-term
mean Theta for regime r;
Sigma: a k x k x R array, each Sigma[,, r] containing the
matrix Sigma for regime r;
Sigmae: a k x k x R array, each Sigmae[,,r] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r
Sigmaj: a k x k x R array, each Sigmaj[,,r] containing the
matrix Sigmaj for regime r;
mj: a k x R matrix, row mj[, r] containing mj for regime r;
xi: a M-1 binary vector indicating jump or not}

\item{pc}{a k x M logical matrix representing the present coordinates at each
node}
}
\value{
a named list containing the following elements:
A: a k x k x M array, A[,,i] corresponding to Ai for
each branch ending at node or tip i;
b: a k x M matrix, b[,i] corresponding to the vectors bi;
C: a k x k x M array, C[,,i] corresponding to the
matrices Ci;
d: a k x M matrix, d[,i] corresponding to the vectors di;
E: a k x k x M array, E[,,i] corresponding to the matrices Ei;
f: a vector, f[i] correspondign to fi
}
\description{
Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree
}
\details{
The dimnames
}
