% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOU.R
\name{AbCdEf.OU}
\alias{AbCdEf.OU}
\title{Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree}
\usage{
AbCdEf.OU(params, metaI = validateParameters.OU(params, verbose = verbose),
  pc, verbose = FALSE)
}
\arguments{
\item{params}{parameters of the OU process. This must be a
named list with the following elements:
A: a R x k x k array, where R is the number of regimes of the
OU process, k is the number of variables (traits), each A[r,,]
containing the matrix A for regime r;
Theta: a R x k matrix, row Theta[r, ] containing the long-term
mean theta for regime r;
Sigma: a R x k x k array, each Sigma[r,,] containing the
matrix Sigma for regime r;
Sigmae: a R x k x k array, each Sigmae[r,,] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{tree}{a phylo object (see details)}

\item{presentCoords}{a M x k logical matrix representing the present coordinates at each
node}
}
\value{
a named list containing the following elements:
A: a M x k x k array, A[i,,] corresponding to Ai for
each branch ending at node or tip i;
b: a M x k matrix, b[i,] corresponding to the vectors bi;
C: a M x k x k array, C[i,,] corresponding to the
matrices Ci;
d: a M x k matrix, d[i,] corresponding to the vectors di;
E: a M x k x k array, E[i,,] corresponding to the matrices Ei;
f: a vector, f[i] correspondign to fi
}
\description{
Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree
}
\details{
The dimnames
}

