% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{AbCdEf}
\alias{AbCdEf}
\title{Quadratic polynomial parameters A, b, C, d, E, f for each node}
\usage{
AbCdEf(tree, model, metaI, pc, verbose = FALSE)
}
\description{
A generic function specified for every model class. This
function is called by mvlik.
}
