% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{AbCdEf.TwoSpeedOU}
\alias{AbCdEf.TwoSpeedOU}
\title{Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree}
\usage{
\method{AbCdEf}{TwoSpeedOU}(tree, model,
  metaI = validateModel.TwoSpeedOU(tree, model, verbose = verbose), pc,
  verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object (see details)}

\item{model}{parameters of the TwoSpeedOU process. This must be a
named list with the following elements:
Alpha1, Alpha2: k x k x R arrays, where R is the number of regimes of the
TwoSpeedOU process, k is the number of variables (traits)
Theta: a k x R matrix, row Theta[, r] containing the long-term
mean Theta for regime r;
Sigma: a k x k x R array, each Sigma[,,r] containing the
matrix Sigma for regime r;
Sigmae: a k x k x R array, each Sigmae[,,r] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{presentCoords}{a M x k logical matrix representing the present coordinates at each
node}
}
\value{
a named list containing the following elements:
A: a k x k x M array, A[,,i] corresponding to Ai for
each branch ending at node or tip i;
b: a k x M matrix, b[,i] corresponding to the vectors bi;
C: a k x k x M array, C[,,i] corresponding to the
matrices Ci;
d: a k x M matrix, d[,i] corresponding to the vectors di;
E: a k x k x M array, E[,,i] corresponding to the matrices Ei;
f: a vector, f[i] correspondign to fi
}
\description{
Calculate the coefficients A, b, C, d, E, f of the general
form (eq. 1) for each edge in a tree
}
\details{
The dimnames
}
