% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{Args_MixedGaussian_MGPMScalarOUType}
\alias{Args_MixedGaussian_MGPMScalarOUType}
\title{Arguments for the MixedGaussian constructor for scalar OU MGPM models.}
\usage{
Args_MixedGaussian_MGPMScalarOUType()
}
\value{
a list.
}
\description{
Arguments for the MixedGaussian constructor for scalar OU MGPM models.
}
