% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{Args_MixedGaussian_MGPMSurfaceOUType}
\alias{Args_MixedGaussian_MGPMSurfaceOUType}
\title{Arguments for the MixedGaussian constructor for SURFACE OU MGPM models.}
\usage{
Args_MixedGaussian_MGPMSurfaceOUType()
}
\value{
a list.
}
\description{
Arguments for the MixedGaussian constructor for SURFACE OU MGPM models.
}
