% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{as.MGPM}
\alias{as.MGPM}
\title{If necessary, convert an object to an MGPM object.
This is an S3 generic function.}
\usage{
as.MGPM(o, ctx = NULL)
}
\arguments{
\item{o}{an object.}

\item{ctx}{a \code{\link{MGPMContext}} object. Default \code{NULL}.}
}
\value{
an MGPM object corresponding to \code{o}. If \code{o} is already
a \code{\link{MGPM}} object, it is returned as is. Otherwise, a
\code{\link{MGPM}} object is constructed from \code{o} and,
optionally \code{ctx}.
}
\description{
If necessary, convert an object to an MGPM object.
This is an S3 generic function.
}
