% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{as.MGPMVector}
\alias{as.MGPMVector}
\title{Convert a MGPM object to a MGPMVector object}
\usage{
as.MGPMVector(o)
}
\arguments{
\item{o}{a MGPMVector or a MGPM object}
}
\value{
a MGPMVector object corresponding to o.
}
\description{
Convert a MGPM object to a MGPMVector object
}
\seealso{
\code{\link{MGPM}} \code{\link{MGPMVector}}
}
