% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTable.R
\name{FormatAsLatex}
\alias{FormatAsLatex}
\title{Latex represenataion of an object.}
\usage{
FormatAsLatex(x, ...)
}
\arguments{
\item{x}{an R object. Currently, objects of S3 classes MatrixParameter,
VectorParameter, ScalarParameter and PCMTable are supported.}

\item{...}{additional arguments passed from calling functions.}
}
\description{
Latex represenataion of an object.
}
