% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTable.R
\name{FormatCellAsLatex}
\alias{FormatCellAsLatex}
\title{Latex represenataion of a model parameter or other found in a PCMTable object}
\usage{
FormatCellAsLatex(x)
}
\arguments{
\item{x}{an R object. Currently, objects of S3 classes MatrixParameter,
VectorParameter, ScalarParameter and PCMTable are supported.}
}
\value{
a character string
}
\description{
Latex represenataion of a model parameter or other found in a PCMTable object
}
