% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatexTools.R
\name{FormatTableAsLatex}
\alias{FormatTableAsLatex}
\title{Latex representation of a data.table with matrix and vectors in its cells}
\usage{
FormatTableAsLatex(x, argsXtable = list(), ...)
}
\arguments{
\item{x}{a data.table}

\item{argsXtable}{a list (empty list by default) passed to xtable...}

\item{...}{additional arguments passed to print.xtable.}
}
\value{
a character string representing a parseable latex text.
}
\description{
Latex representation of a data.table with matrix and vectors in its cells
}
\examples{
dt <- data.table::data.table(
   A = list(
         matrix(c(2, 0, 1.2, 3), 2, 2),
         matrix(c(2.1, 0, 1.2, 3.2, 1.3, 3.4), 3, 2)),
   b = c(2.2, 3.1))
print(FormatTableAsLatex(dt))

}
