% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{GenerateDefaultParameterizations}
\alias{GenerateDefaultParameterizations}
\title{This function was used to autogenerate the code for PCMSpecify.OU, .JOU, .BM, .White and .DOU}
\usage{
GenerateDefaultParameterizations(libname, pkgname)
}
\description{
This function was used to autogenerate the code for PCMSpecify.OU, .JOU, .BM, .White and .DOU
}
