% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOU.R
\name{generateMVTrait.OU}
\alias{generateMVTrait.OU}
\title{Generate multivariate trait-values on a tree according to a multivariate OU
process with one or several regimes}
\usage{
generateMVTrait.OU(root.value, params, verbose = FALSE)
}
\arguments{
\item{root.value}{numeric vector of length k indicating the root value}

\item{tree}{a phylo object}

\item{A}{numeric k x k matrix specifying the selection strengths of the OU
process}

\item{Theta}{numeric k-vector specifying the long-term mean of the OU process}

\item{Sigma}{numeric k x k matrix specifying the stochastic time-unit variance of the
OU process;}

\item{A, Theta, Sigma}{parameters of the OU process:
A: a R x k x k array, where R is the number of regimes of the
OU process, k is the number of variables (traits), each A[r,,]
containing the matrix A for regime r;
Theta: a R x k matrix, row Theta[r, ] containing the long-term
mean theta for regime r;
Sigma: a R x k x k array, each Sigma[r,,] containing the
matrix Sigma for regime r;
Sigmae: a R x k matrix, each Sigmae[r, ] containing the environmental
variances for the k traits in regime r}
}
\value{
a numeric M x k matrix of values at all nodes of the tree (root,
internal and tip), where M is the number such nodes: M=dim(tree$edge)[1]+1,
with indices from 1 to N=length(tree$tip.label) corresponding to tips, N+1
corresponding to the root and bigger than N+1 corresponding to internal nodes.
}
\description{
Generate multivariate trait-values on a tree according to a multivariate OU
process with one or several regimes
}

