% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianPCM.R
\name{is.GaussianPCM}
\alias{is.GaussianPCM}
\title{Check if an object is a `GaussianPCM`}
\usage{
is.GaussianPCM(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE if x inherits from the S3 class `GaussianPCM`, FALSE otherwise.
}
\description{
Check if an object is a `GaussianPCM`
}
