% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{is.MGPMContext}
\alias{is.MGPMContext}
\title{Check if an object is of S3 class 'MGPMContext'}
\usage{
is.MGPMContext(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MGPMContext'
}
