% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{is.MGPMVector}
\alias{is.MGPMVector}
\title{Check if an object is of S3 class 'MGPMVector'}
\usage{
is.MGPMVector(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MGPMVector'
}
