% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedGaussian.R
\name{is.MixedGaussian}
\alias{is.MixedGaussian}
\title{Check if an object is a `MixedGaussian` PCM}
\usage{
is.MixedGaussian(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE if x inherits from the S3 class `MixedGaussian`, FALSE otherwise.
}
\description{
Check if an object is a `MixedGaussian` PCM
}
