% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{is.PCM}
\alias{is.PCM}
\title{Check if an object is a PCM.}
\usage{
is.PCM(x)
}
\arguments{
\item{x}{an object.}
}
\value{
TRUE if `x` inherits from the S3 class "PCM".
}
\description{
Check if an object is a PCM.
}
