% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{is.PCMTree}
\alias{is.PCMTree}
\title{Check that a tree is a PCMTree}
\usage{
is.PCMTree(tree)
}
\arguments{
\item{tree}{a tree object.}
}
\value{
a logical TRUE if `inherits(tree, "PCMTree")` is TRUE.
}
\description{
Check that a tree is a PCMTree
}
