% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{Lmr.default}
\alias{Lmr.default}
\title{Defaut R-implementation of Lmr}
\usage{
\method{Lmr}{default}(X, tree, model, metaI = validateModel(tree, model),
  pruneI = pruneTree(tree), pc = presentCoordinates(X, tree, pruneI),
  root.only = FALSE, verbose = FALSE)
}
\value{
A list with the members A,b,C,d,E,f,L,m,r for all nodes in the tree.
}
\description{
Defaut R-implementation of Lmr
}
\details{
This funciton is not a generic S3 implementation of Lmr - it needs
additional raguments and is designed to be called by mvlik. It can still
be called by the end-user for debugging purpose.
}
