% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{Lmr}
\alias{Lmr}
\title{Quadratic polynomial parameters L, m, r}
\usage{
Lmr(X, tree, model, metaI = validateModel(tree, model, verbose = verbose),
  pruneI = pruneTree(tree), pc = presentCoordinates(X, tree, pruneI),
  root.only = TRUE, verbose = FALSE)
}
\description{
Quadratic polynomial parameters L, m, r
}
