% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{MGPMContext}
\alias{MGPMContext}
\title{Context of a MGPM}
\usage{
MGPMContext(X, tree, model, SE = matrix(0, PCMNumTraits(model),
  PCMTreeNumTips(tree)), tipModelType = NULL)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} and \code{NaN} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Missing values can be either not-available (\code{NA}) or not existing (\code{NaN}).
These two values have are treated differently when calculating
likelihoods: see \code{\link{PCMPresentCoordinates}}.}

\item{tree}{a phylo object with N tips.}

\item{model}{a MixedGaussian model used as a template to build MixedGaussian
model objects.}

\item{SE}{a k x N matrix specifying the standard error for each measurement in
X. Alternatively, a k x k x N cube specifying an upper triangular k x k
factor of the variance covariance matrix for the measurement error
for each node i=1, ..., N.
Default: \code{matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree))}.}

\item{tipModelType}{NULL or a character string indicating a member of
\code{model}. This argument allows to set a specific regime for all terminal
branches, such as a white noise model.}
}
\value{
an object of S3 class 'MGPMContext'.
}
\description{
Context of a MGPM
}
