% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{MGPMPosK}
\alias{MGPMPosK}
\alias{MGPMPosn}
\alias{MGPMPosl}
\alias{MGPMPosr}
\alias{MGPMPosm}
\alias{MGPMPosv}
\title{Indices of different parts of an MGPM in an MGPMVector}
\usage{
MGPMPosK(s, ctx = NULL)

MGPMPosn(s, ctx = NULL)

MGPMPosl(s, ctx = NULL)

MGPMPosr(s, ctx = NULL)

MGPMPosm(s, ctx = NULL)

MGPMPosv(s, ctx = NULL)
}
\arguments{
\item{s}{a \code{\link{MGPM}} or a \code{\link{MGPMVector}}
object.}

\item{ctx}{a \code{\link{MGPMContext}} object. This is needed only if
\code{s} is \code{\link{MGPMVector}} object.}
}
\value{
an integer vector.
}
\description{
The different parts of an MGPM are described in
\code{\link{MGPMVector}}. \code{posK} returns the osition of the
 number of shifts (this is always equal to 1). See the section functions
 for the others.
}
\section{Functions}{
\itemize{
\item \code{MGPMPosn}: Positions of the shift nodes;

\item \code{MGPMPosl}: Positions of the offsets of the shift-points in tip-ward direction
relative to the beginnings of the branches leading to shift nodes;

\item \code{MGPMPosr}: Position of the regime id's for each shift node;

\item \code{MGPMPosm}: Positions of the model type id's for each regime;

\item \code{MGPMPosv}: Positions of the model parameters.
}}

\seealso{
\code{\link{MGPM}}, \code{\link{MGPMVector}}
}
