% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{MGPMScalarOUType}
\alias{MGPMScalarOUType}
\title{Class name for the scalar OU MGPM model type}
\usage{
MGPMScalarOUType()
}
\value{
a character vector of one named element (ScalarOU)
}
\description{
Class name for the scalar OU MGPM model type
}
