% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{MGPMSurfaceOUType}
\alias{MGPMSurfaceOUType}
\title{Class name for the SURFACE OU MGPM model type}
\usage{
MGPMSurfaceOUType()
}
\value{
a character vector of one named element (SURFACE)
}
\description{
Class name for the SURFACE OU MGPM model type
}
