% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{MGPMVector}
\alias{MGPMVector}
\title{Construct a MGPM vector.}
\usage{
MGPMVector(K, n, l, r, m, v)
}
\arguments{
\item{K}{an integer denoting the number of shifts.}

\item{n}{an integer vector of length \code{K} denoting shift nodes.}

\item{l}{a real vector of length \code{K} denoting shift-point offsets in
tip-ward direction from the beginning of the branches leading to shift-nodes.}

\item{r}{an integer vector of length \code{K} denoting the regime
associated with each part (see Details in \code{\link{MGPM}}).}

\item{m}{an integer vector of length \code{R}, where \code{R} denotes the
total number of unique regimes (see Details in \code{\link{MGPM}}).}

\item{v}{a real vector of length \code{P}, where \code{P} denotes the total
number of numeric parameters of the MixedGaussian model (see Details in
\code{\link{MGPM}}).}
}
\value{
a real (double) vector representing the concatenation of
\code{K, n, l, r, m, v}, with S3 class set to 'MGPMVector'.
}
\description{
Construct a MGPM vector.
}
\seealso{
\code{\link{MGPM}}
}
