% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRG.R
\name{MRG}
\alias{MRG}
\alias{is.MRG}
\alias{MRG}
\title{Multiple regime Gaussian PCMs}
\usage{
is.MRG(x)

MRG(k, modelTypes, mapping, className = paste0("MRG_", do.call(paste0,
  as.list(mapping))), X0 = list(default = rep(0, k), type = c("gvector",
  "full"), description = "trait vector at the root; global for all regimes"),
  Sigmae_x = list(default = matrix(0, nrow = k, ncol = k), type = c("gmatrix",
  "upper.tri.diag", "positive.diag"), description =
  "Upper triangular Choleski factor of the variance-covariance matrix for the non-phylogenetic trait component"),
  ...)
}
\arguments{
\item{k}{integer defining the number of traits.}

\item{modelTypes}{a character string vector with the class names of the
model-types that can possibly be included (assigned to regimes) in the MRG,
e.g. c("BM3", "OU3").}

\item{mapping}{a character string vector with elements from modelTypes or an
integer vector with elements between 1 and length(modelTypes)
mapping modelTypes to regimes, e.g. if \code{modelTypes = c("BM3", "OU3")} and
\code{mapping = c(a = 1, b = 1, c = 2, d = 1)} defines an MRG with four
different regimes with model-types BM3, BM3, OU3 and BM3, corresponding to each
regime. \code{mapping} does not have to be a named vector. If it is a named
vector, then all the names must correspond to valid regime names in a tree to
which the model will be fit (member tree$edge.regime should be a character vector).
If it is not a named vector than the positions of the elements correspond to the
regimes in their order given by the function \code{\link{PCMTreeUniqueRegimes}}
called on a tree object.}

\item{className}{a character string definingn a valid S3 class name for the
resulting MRG object. If not specified, a className is generated using the
expression \code{ paste0("MRG_", do.call(paste0, as.list(mapping)))}.}

\item{X0}{NULL or a list defining a global vector X0 to be used by all
models in the MRG.}

\item{Sigmae_x}{NULL or a list defining a global matrix to be used as element
Sigmae_x by all models in the MRG.}
}
\value{
an object of S3 class className inheriting from MRG, GaussianPCM and
PCM.
}
\description{
Create a multi-regime Gaussian model (MRG)
}
\details{
If X0 is not NULL it has no sense to use model-types including X0 as a
parameter (e.g. use BM1 or BM3 insted of BM or BM2). Similarly if Sigmae_x is
not NULL there is no meaning in using model-types including Sigmae_x as a parameter,
(e.g. use OU2 or OU3 instead of OU or OU1).
}
\seealso{
\code{\link{PCMTreeUniqueRegimes}}
}
