% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{mvcond.BM}
\alias{mvcond.BM}
\title{Create a conditional multivariate BM distribution}
\usage{
\method{mvcond}{BM}(tree, model, r = 1, verbose = FALSE)
}
\arguments{
\item{Sigma}{of the multivariate BM process; Sigma is a k x k matrix}
}
\value{
a list containging the passed parameters as well as
a function mvr of arguments n (number of observation k-vectors to generate),
x0 (initial k-vector of values), t (numeric time); and a function mvd for
calculating the density of multivariate vector under the specified distribution
and given an initial value and time.
}
\description{
Create a conditional multivariate BM distribution
}
