% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{mvcond.JOU}
\alias{mvcond.JOU}
\title{Create a conditional multivariate JOU distribution}
\usage{
\method{mvcond}{JOU}(tree, model, r = 1, verbose = FALSE)
}
\arguments{
\item{H, Theta, Sigma, Sigmaj, mj, xi}{parameters of the multivariate JOU process; H is a k x k
matrix, Theta,mj are k-vectors and Sigma,Sigmaj are k x k matrices,xi is a vector of length equal to the number
of edges in the tree}
}
\value{
a list containging the passed parameters as well as
a function mvr of arguments n (number of observation k-vectors to generate),
x0 (initial k-vector of values), t (numeric time); and a function mvd for
calculating the density of multivariate vector under the specified distribution
and given an initial value and time.
}
\description{
Create a conditional multivariate JOU distribution
}
