% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOU.R
\name{mvCond.OU}
\alias{mvCond.OU}
\title{Create a conditional multivariate OU distribution}
\usage{
mvCond.OU(A, theta, Sigma)
}
\arguments{
\item{A, theta, Sigma}{parameters of the multivariate OU process; A is a k x k
matrix, theta is a k-vector and Sigma is a k x k matrix}
}
\value{
a list containging the passed parameters as well as
a function rmv of arguments n (number of observation k-vectors to generate),
x0 (initial k-vector of values), t (numeric time); and a function dmv for
calculating the density of multivariate vector under the specified distribution
and given an initial value and time.
}
\description{
Create a conditional multivariate OU distribution
}

