% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{mvcond.TwoSpeedOU}
\alias{mvcond.TwoSpeedOU}
\title{Create a conditional multivariate TwoSpeedOU distribution}
\usage{
\method{mvcond}{TwoSpeedOU}(tree, model, r = 1, verbose = FALSE)
}
\arguments{
\item{Alpha1, Alpha2, Theta, Sigma}{parameters of the multivariate TwoSpeedOU process; Alpha1,Alpha2 are
k x k matrices, Theta is a k-vector and Sigma is a k x k matrix}
}
\value{
a list containging the passed parameters as well as
a function mvr of arguments n (number of observation k-vectors to generate),
x0 (initial k-vector of values), t (numeric time); and a function mvd for
calculating the density of multivariate vector under the specified distribution
and given an initial value and time.
}
\description{
Create a conditional multivariate TwoSpeedOU distribution
}
