% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{mvlik}
\alias{mvlik}
\title{Multivariate likelihood calculation}
\usage{
mvlik(X, tree, model, metaI = validateModel(tree, model, verbose = verbose),
  pruneI = pruneTree(tree), pc = presentCoordinates(X, tree), log = TRUE,
  verbose = FALSE, debug = FALSE)
}
\description{
Multivariate likelihood calculation
}
