% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOU.R
\name{orderBreadthFirst}
\alias{orderBreadthFirst}
\title{Breadth-first tree traversal}
\usage{
orderBreadthFirst(tree)
}
\arguments{
\item{tree}{a phylo object with possible singleton nodes (i.e. internal nodes with
one daughter node)}
}
\value{
a vector of indices of edges in tree$edge in breadth-first order.
}
\description{
Breadth-first tree traversal
}

