% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{parseErrorMessage}
\alias{parseErrorMessage}
\title{Extract error information from a formatted error message.}
\usage{
parseErrorMessage(x)
}
\arguments{
\item{x}{character string representing the error message.}
}
\value{
a named list with the parsed error information or NULL, if no match
was found.
}
\description{
The function searches x for a pattern matching the format
'ERR:5-digit-code:project-name:source-file:error-specifics:'. Specifically it
searches for a regular expression pattern "ERR:[0-9]+:[^:]+:[^:]+:[^:]+:[^:]*:".
}
