% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCM}
\alias{PCM}
\title{Create a phylogenetic comparative model object}
\usage{
PCM(model, k = 1, regimes = 1, params = NULL, vecParams = NULL,
  offset = 0, ...)
}
\arguments{
\item{model}{This argument can take one of the following forms:
\itemize{
\item a character vector of the S3-classes of the model object to be
created (one model object can have one or more S3-classes, with the class
PCM at the origin of the hierarchy);
\item an S3 object which's class inherits from the PCM S3 class.
}
The Details section explains how these two types of input are processed.}

\item{k}{integer denoting the number of traits (defaults to 1).}

\item{regimes}{a character or integer vector denoting the regimes.}

\item{params}{NULL (default) or a list of parameter values (scalars, vectors,
matrices, or arrays) or sub-models (S3 objects inheriting from the PCM class).
See details.}

\item{vecParams}{NULL (default) or a numeric vector the vector
representation of the variable parameters in the model. See details.}

\item{offset}{integer offset in vecParams; see Details.}

\item{...}{additional parameters intended for use by sub-classes of the PCM
class.}
}
\value{
an object of S3 class as defined by the argument model.
}
\description{
This is the entry-point function for creating model objects within
the PCMBase framework. A PCM
}
