% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMBaseIsADevRelease}
\alias{PCMBaseIsADevRelease}
\title{Check if the PCMBase version corresponds to a dev release}
\usage{
PCMBaseIsADevRelease()
}
\value{
a logical
}
\description{
Check if the PCMBase version corresponds to a dev release
}
