% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMCombineListAttribute}
\alias{PCMCombineListAttribute}
\title{Combine all member attributes of a given name into a list}
\usage{
PCMCombineListAttribute(object, name)
}
\arguments{
\item{object}{a named list object.}

\item{name}{a character string denoting the name of the attribute.}
}
\value{
a list of attribute values
}
\description{
Combine all member attributes of a given name into a list
}
