% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU.R
\name{PCMCond.OU}
\alias{PCMCond.OU}
\title{Create a conditional multivariate OU distribution}
\usage{
\method{PCMCond}{OU}(tree, model, r = 1, verbose = FALSE)
}
\arguments{
\item{H, Theta, Sigma}{parameters of the multivariate OU process; H is a k x k
matrix, Theta is a k-vector and Sigma is a k x k matrix}
}
\value{
a list containging the passed parameters as well as
a function `random` of arguments n (number of observation k-vectors to generate),
x0 (initial k-vector of values), t (numeric time); and a function `density` for
calculating the density of multivariate vector under the specified distribution
and given an initial value and time.
}
\description{
Create a conditional multivariate OU distribution
}
