% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMDefaultObject}
\alias{PCMDefaultObject}
\title{Generate a default object of a given PCM model type or parameter type}
\usage{
PCMDefaultObject(spec, model, ...)
}
\arguments{
\item{spec}{any object having a class attribute. The value of this object is not
used, but its class is used for method-dispatch.}

\item{model}{a PCM object used to extract attributes needed for creating a
default object of class specified in \code{class(spec)}, such as the number of
traits (k) or the regimes and the number of regimes;}
}
\value{
a parameter or a PCM object.
}
\description{
Generate a default object of a given PCM model type or parameter type
}
