% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMDescribe}
\alias{PCMDescribe}
\title{Human friendly description of a PCM}
\usage{
PCMDescribe(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a character string
}
\description{
Human friendly description of a PCM
}
\details{
This S3 generic function is intended to be specified for user models
}
