% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMDescribeParameters}
\alias{PCMDescribeParameters}
\title{Describe the parameters of a PCM}
\usage{
PCMDescribeParameters(model, ...)
}
\arguments{
\item{model}{a PCM object.}

\item{...}{additional arguments that can be used by implementing methods.}
}
\value{
a named list with character elements corresponding to each parameter.
}
\description{
This is an S3 generic.
}
