% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMExtractRegimes}
\alias{PCMExtractRegimes}
\title{Given a PCM or a parameter object, extract an analogical object for a subset
of the regimes in the original object.}
\usage{
PCMExtractRegimes(obj, regimes = seq_len(PCMNumRegimes(obj)))
}
\arguments{
\item{obj}{a PCM or a parameter object.}

\item{regimes}{an integer vector; should be a subset or equal to
\code{seq_len(PCMNumRegimes(obj))} (the default).}
}
\value{
an object of the same class as obj with a subset of obj's regimes
}
\description{
Given a PCM or a parameter object, extract an analogical object for a subset
of the regimes in the original object.
}
\details{
This is an S3 generic
}
