% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMFindMethod}
\alias{PCMFindMethod}
\title{Find the S3 method for a given PCM object or class-name and an S3 generic}
\usage{
PCMFindMethod(x, method = "PCMCond")
}
\arguments{
\item{x}{a character string denoting a PCM S3 class name (e.g. "OU"), or a
PCM object.}

\item{method}{a character string denoting the name of an S3 generic function.
Default: "PCMCond".}
}
\value{
a function object corresponding to the S3 method found or an error is
raised if no such function is found for the specified object and method.
}
\description{
Find the S3 method for a given PCM object or class-name and an S3 generic
}
