% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMFit}
\alias{PCMFit}
\title{Fitting a PCM model to a given tree and data}
\usage{
PCMFit(X, tree, model, metaI = PCMInfo(X, tree, model), ...)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Depending on the value of the global option "PCMBase.Internal.PC.Full" \code{NA}
entries are interpreted either as missing measurements or non-existing traits
(see \code{\link{PCMPresentCoordinates}}).}

\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{metaI}{a list returned from a call to \code{PCMInfo(X, tree, model)},
containing meta-data such as N, M and k.}
}
\value{
an object of class PCMFit
}
\description{
Fitting a PCM model to a given tree and data
}
