% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMFixParameter}
\alias{PCMFixParameter}
\title{Fix a parameter in a PCM model}
\usage{
PCMFixParameter(model, name)
}
\arguments{
\item{model}{a PCM object}

\item{name}{a character string}
}
\value{
a copy of the model with added class '_Fixed' to the class of the
parameter \code{name}
}
\description{
Fix a parameter in a PCM model
}
