% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMGetVecParams}
\alias{PCMGetVecParams}
\title{Get a vector of the varying (non-fixed) parameters in a PCM model}
\usage{
PCMGetVecParams(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a numeric vector of length PCMNumParams(model).
}
\description{
Get a vector of the varying (non-fixed) parameters in a PCM model
}
\seealso{
\code{\link{PCMSetOrGetVecParams}} \code{\link{PCMGetVecParamsFull}}
}
