% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMGetVecParamsFull}
\alias{PCMGetVecParamsFull}
\title{Get a vector with all model parameters unrolled}
\usage{
PCMGetVecParamsFull(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a numerical vector
}
\description{
Get a vector with all model parameters unrolled
}
