% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMGetVecParamsRegimesAndModels}
\alias{PCMGetVecParamsRegimesAndModels}
\title{Get a vector of all parameters (real and discrete) describing a model on a
tree including the numerical parameters of each model regime, the integer ids
of the splitting nodes defining the regimes on the tree and the integer ids of
the model types associated with each regime.}
\usage{
PCMGetVecParamsRegimesAndModels(model, tree, ...)
}
\arguments{
\item{model}{a PCM model}

\item{tree}{a phylo object with an edge.part member.}

\item{...}{additional parameters passed to methods.}
}
\value{
a numeric vector concatenating the result
}
\description{
Get a vector of all parameters (real and discrete) describing a model on a
tree including the numerical parameters of each model regime, the integer ids
of the splitting nodes defining the regimes on the tree and the integer ids of
the model types associated with each regime.
}
\details{
This is an S3 generic.
In the default implementation, the last entry in the returned vector is the
number of numerical parameters. This is used to identify the starting positions
in the vector of the first splitting node.
}
