% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMJumps}
\alias{PCMJumps}
\title{Jumps in modeled traits associated with branches in a tree}
\usage{
PCMJumps(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
an integer vector of 0's and 1's with entries correspondin to the
denoting if a jump took place at the beginning of a branch.
}
\description{
Jumps in modeled traits associated with branches in a tree
}
