% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMListParameterizations}
\alias{PCMListParameterizations}
\title{Specify the parameterizations for each parameter of a model}
\usage{
PCMListParameterizations(model, ...)
}
\arguments{
\item{model}{a PCM.}

\item{...}{additional arguments used by implementing methods.}
}
\value{
a named list with list elements corresponding to each parameter in
model. Each list element is a list of character vectors, specifying the possible
S3 class attributes for the parameter in question. For an example, type
`PCMListParameterizations.BM` to see the possible parameterizations for the
BM model.
}
\description{
This is an S3 generic.
}
