% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMLmr}
\alias{PCMLmr}
\title{Quadratic polynomial parameters L, m, r}
\usage{
PCMLmr(X, tree, model, metaI = PCMInfo(X, tree, model, verbose = verbose),
  root.only = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Depending on the value of the global option "PCMBase.Internal.PC.Full" \code{NA}
entries are interpreted either as missing measurements or non-existing traits
(see \code{\link{PCMPresentCoordinates}}).}

\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{metaI}{a list returned from a call to \code{PCMValidate(tree, model)},
containing meta-data such as N, M and k.}

\item{root.only}{logical indicatin whether to return the calculated values of L,m,r
only for the root or for all nodes in the tree.}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\value{
A list with the members A,b,C,d,E,f,L,m,r for all nodes in the tree or
  only for the root if root.only=TRUE.
}
\description{

}
