% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMLoadMatrixParameter}
\alias{PCMLoadMatrixParameter}
\title{Load/store a matrix parameter from/to a vector of all parameters in a model.}
\usage{
PCMLoadMatrixParameter(m, vecParams, offset, k, type = c("diag1", "diag",
  "upper.tri", "upper.tri.diag", "lower.tri", "lower.tri.diag", "symmetric",
  "full", "custom", "fixed"), mask = matrix(TRUE, k, k),
  indices = function(offset, k) offset + (1:(k * k)), load = TRUE)
}
\arguments{
\item{m}{numeric k x k matrix. If load==TRUE (default) this has to be a matrix-object in
the calling environment (not a temporary object such as the result from an algebraic expression).}

\item{vecParams}{numeric vector containing all parameters in the model. If load==FALSE, this
has to be a vector in the calling environment (not a temporary object such as the result from
an algebraic expression).}

\item{offset}{integer denoting the offset in vecParams (0 means start from vecParams[1] onwards).}

\item{k}{integer denoting the dimensionality (number of rows and colums) of the resulting matrix.}

\item{type}{a character string indicating the type of loading. possible values are:
\describe{
\item{"diag1"}{create a diagonal matrix with all diagonal elements repeating \code{vecParams[offset+1]}.}
\item{"diag"}{create a diagonal matrix with a diagonal copied from \code{vecParams[offset + (1:k)]}.}
\item{"upper.tri"}{load an upper triangular matrix with zero diagonal.}
\item{"upper.tri.diag"}{load an upper triangular matrix including diagonal.}
\item{"lower.tri"}{load a lower triangular matrix with zero diagonal.}
\item{"lower.tri.diag"}{load a lower triangular matrix including diagonal.}
\item{"symmetric"}{load a symmetric matrix. Only the elements of the upper triangle and diagonal are
loaded from vecParams.}
\item{"full"}{load a full.}
\item{"custom"}{use a custom template matrix which's elements specified by mask are
to be replaced with \code{vecParams[offset + indices]}}.
}}

\item{mask}{a logical k x k matrix pointing to elements in template to be replaced. used only with type=="custom".}

\item{indices}{a function of the form function(offset, k) returning an integer vector of
length equal to the number of TRUE elements in mask, indicating the position in vecParams to
be copyed from. used only with type=="custom".}

\item{load}{a logical indicating if loading from or storing to vecParams should be done.}
}
\value{
an integer denoting the number of elemnents read or written from/to vecParams.
In the case of type=="custom",
the number of indices bigger than offset returned by the function indices(offset, k).
}
\description{
Load/store a matrix parameter from/to a vector of all parameters in a model.
}
\details{
This function has both, a returned value and side effects. By default the function
loads elements from vecParams into m. This behavior is reversed if the argument load
is set to FALSE.
}
