% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMLoadVectorParameter}
\alias{PCMLoadVectorParameter}
\title{Load/store a vector parameter from/to a vector of all parameters in a model.}
\usage{
PCMLoadVectorParameter(v, vecParams, offset, k, type = c("rep", "full",
  "custom"), mask = rep(TRUE, k), indices = function(offset, k) offset +
  (1:k), load = TRUE)
}
\arguments{
\item{v}{numeric k-vector. If load==TRUE (default) this has to be a vector-object in
the calling environment (not a temporary object such as the result from an algebraic expression).}

\item{vecParams}{numeric vector containing all parameters in the model. If load==FALSE, this
has to be a vector in the calling environment (not a temporary object such as the result from
an algebraic expression).}

\item{offset}{integer denoting the offset in vecParams (0 means start from vecParams[1] onwards).}

\item{k}{integer denoting the dimensionality (length) of the resulting vector.}

\item{type}{a character string indicating the type of loading. possible values are:
\describe{
\item{"rep"}{repeat \code{vecParams[offset+1]} k times}
\item{"full"}{copy \code{vecParams[offset + (1:k)]}}
\item{"custom"}{use a custom template vector which elements specified by mask are
to be replaced with \code{vecParams[indices(offset, k)]}}
}.}

\item{mask}{a logical vecgtor pointing to elements in template to be replaced. used only with type=="custom".}

\item{indices}{a function of the form function(offset, k) returning an integer vector of
length equal to the number of TRUE elements in mask, indicating the position in vecParams to
be copyed from. used only with type=="custom".}

\item{load}{a logical indicating if loading from or storing to vecParams should be done.}
}
\value{
an integer denoting the number of elemnents read or written from/to vecParams.
In the case of type=="custom",
the number of indices bigger than offset returned by the function indices(offset, k).
}
\description{
Load/store a vector parameter from/to a vector of all parameters in a model.
}
\details{
This function has both, a returned value and side effects. By default the function
loads elements from vecParams into v. This behavior is reversed if the argument load
is set to FALSE.
}
