% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMLowerBound}
\alias{PCMLowerBound}
\title{Numerical lower bound}
\usage{
PCMLowerBound(model, ...)
}
\arguments{
\item{model}{a PCM object}
}
\value{
a PCM object of the same S3 classes as model. Calling
\code{\link{PCMBase::PCMSetOrGetVecParams}} on this object returns a lower
bound for that can be used, e.g. in a call to \code{\link{optim}}
}
\description{
Numerical lower bound
}
\examples{
model <- PCM("BM3", 3)
PCMLowerBound(model)
}
