% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMMatchRegimesModelTree}
\alias{PCMMatchRegimesModelTree}
\title{Regimes associated with branches in a tree}
\usage{
PCMMatchRegimesModelTree(tree, model)
}
\arguments{
\item{tree}{a phylo object}

\item{model}{a PCM object}
}
\value{
an integer vector with entries corresponding to the rows in tree$edge
  denoting the regime on each branch in the tree
}
\description{
Regimes associated with branches in a tree
}
