% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMModels}
\alias{PCMModels}
\title{Get a list of PCM models currently implemented}
\usage{
PCMModels(pattern = NULL, parentClass = NULL, ...)
}
\arguments{
\item{pattern}{a character string specifying an optional for the model-names to search for.}

\item{parentClass}{a character string specifying an optional parent class of the models to look for.}

\item{...}{additional arguments used by implementing methods.}
}
\value{
a character vector of the model classes found.
}
\description{
Get a list of PCM models currently implemented
}
\details{
The function is using the S3 api function \code{\link{methods}} looking
for all registered implementations of the function \code{\link{PCMSpecify}}.
}
