% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMModelTypes}
\alias{PCMModelTypes}
\title{Get the model type(s) of a model}
\usage{
PCMModelTypes(obj)
}
\arguments{
\item{obj}{a PCM object}
}
\value{
a character vector
}
\description{
For a regular PCM object, the model type is its S3 class. For a
MixedGaussian each regime is mapped to one of several possible model types.
}
