% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMNumNodes}
\alias{PCMNumNodes}
\title{Number of all nodes in a tree}
\usage{
PCMNumNodes(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
the number of nodes in tree including root, internal and tips.
}
\description{
Number of all nodes in a tree
}
\details{
Wrapper for nrow(tree$edge) + 1
}
