% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMNumParams}
\alias{PCMNumParams}
\title{Number of free parameters describing fully a PCM}
\usage{
PCMNumParams(model, ...)
}
\arguments{
\item{model}{a PCM object}

\item{...}{other arguments (possible future use)}
}
\value{
an integer
}
\description{
Number of free parameters describing fully a PCM
}
