% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMNumRegimes}
\alias{PCMNumRegimes}
\title{Number of regimes in a model}
\usage{
PCMNumRegimes(model)
}
\arguments{
\item{model}{a PCM object}
}
\value{
an integer
}
\description{
Number of regimes in a model
}
