% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMNumTraits}
\alias{PCMNumTraits}
\title{Number of traits modeled by a PCM}
\usage{
PCMNumTraits(model)
}
\arguments{
\item{model}{a PCM object}
}
\value{
an integer
}
\description{
Number of traits modeled by a PCM
}
