% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMOptions}
\alias{PCMOptions}
\title{Global options for the PCMBase package}
\usage{
PCMOptions()
}
\value{
a named list with the currently set values of the following global
options:
\itemize{
\item{\code{PCMBase.Value.NA }}{NA value for the likelihood; used in GaussianPCM to
return this value in case of an error occuring
during likelihood calculation. By default, this is set to \code{as.double(NA)}.}
\item{\code{PCMBase.Errors.As.Warnings }}{a logical flag indicating if errors
(occuring, e.g. during likelihood calculation) should be treated as warnings
and added as an attribute "error" to returne likelihood values. Default TRUE.}
\item{\code{PCMBase.Threshold.Lambda_ij }}{a 0-threshold for abs(Lambda_i + Lambda_j),
where Lambda_i and Lambda_j are eigenvalues of the parameter matrix H of an OU or
other model. Default 1e-8. See \code{\link{PCMPExpxMeanExp}}.}
\item{\code{PCMBase.Threshold.SV }}{A 0-threshold for min(svdV)/max(svdV), where
svdV is the vector of eigenvalues of the matrix V for a given branch. The V matrix
is considered singular if it has eigenvalues equal to 0 or when the ratio
min(svdV)/max(svdV) is below PCMBase.Threshold.SV. Default is 1e-6. Treatment
of branches with singular V matrix is defined by the option \code{PCMBase.Singular.Skip}.}
\item{\code{PCMBase.Singular.Skip }}{A logical value indicating whether branches with
singular matrix V should be skipped during likelihood calculation, adding their children
L,m,r values to their parent node. Default TRUE. Note, that setting this option to FALSE
may cause some models to stop working, e.g. the White model. Setting this option to FALSE
will also cause errors or NA likelihood values in the case of trees with very short or
0-length branches.}
\item{\code{PCMBase.Internal.PC.Full }}{A logical value indicating whether NA trait values
are treated as missing measurements which are integrated out during likelihood calculation.
Default TRUE. Setting this option to FALSE will cause NA values to be treated as
non-existing traits and these will be ignored instead of integrated out.}
}
}
\description{
Global options for the PCMBase package
}
