% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMPairSums}
\alias{PCMPairSums}
\title{Sums of pairs of elements in a vector}
\usage{
PCMPairSums(lambda)
}
\arguments{
\item{lambda}{a numeric vector}
}
\value{
a squared symmetric matrix with elem_ij=lambda_i+lambda_j.
}
\description{
Sums of pairs of elements in a vector
}
