% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMParam}
\alias{PCMParam}
\title{Module PCMParam}
\description{
Global and S3 generic functions for manipulating model parameters.
The parameters in a PCM are named objects with a class attribute specifying
the main type and optional properties (tags).

S3 generic functions:
\describe{
\item{PCMParamCount()}{Counting the number of actual numeric parameters (used,
 e.g. for calculating information scores, e.g. AIC);}
\item{PCMParamLoadOrStore(), PCMParamLoadOrStore()}{Storing/loading a parameter
to/from a numerical vector;}
\item{PCMParamLowerLimit(),PCMParamUpperLimit()}{Specifying parameter upper and lower limits;}
\item{PCMParamRandomVecParams()}{Generating a random parameter vector;}
}

For all the above properties, check-functions are defined, e.g. `is.Local(o)`,
`is.Global(o)`, `is.ScalarParameter(o)`, `is.VectorParameter`, etc.
}
