% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMParamUpperLimit}
\alias{PCMParamUpperLimit}
\title{The upper limit for a given model or parameter type}
\usage{
PCMParamUpperLimit(o, k, R, ...)
}
\arguments{
\item{o}{an object such as a VectorParameter a MatrixParameter or a PCM.}

\item{k}{integer denoting the number of traits}

\item{R}{integer denoting the number of regimes in the model in which o
belongs to.}

\item{...}{additional arguments (optional or future use).}
}
\value{
an object of the same S3 class as o representing an upper limit
for the class.
}
\description{
This is an S3 generic function.
}
