% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMParseErrorMessage}
\alias{PCMParseErrorMessage}
\title{Extract error information from a formatted error message.}
\usage{
PCMParseErrorMessage(x)
}
\arguments{
\item{x}{character string representing the error message.}
}
\value{
Currently the function returns \code{x} unchanged.
}
\description{
Extract error information from a formatted error message.
}
