% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMPlotGaussianDensityGrid2D}
\alias{PCMPlotGaussianDensityGrid2D}
\title{A 2D Gaussian distribution density grid in the form of a ggplot object}
\usage{
PCMPlotGaussianDensityGrid2D(
  mu,
  Sigma,
  xlim,
  ylim,
  xNumPoints = 100,
  yNumPoints = 100,
  ...
)
}
\arguments{
\item{mu}{numerical mean vector of length 2}

\item{Sigma}{numerical 2 x 2 covariance matrix}

\item{xlim, ylim}{numerical vectors of length 2}

\item{xNumPoints, yNumPoints}{integers denoting how many points should the grid contain for each axis.}

\item{...}{additional arguments passed to ggplot}
}
\value{
a ggplot object
}
\description{
A 2D Gaussian distribution density grid in the form of a ggplot object
}
