% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMPlotGaussianSample2D}
\alias{PCMPlotGaussianSample2D}
\title{A 2D sample from Gaussian distribution}
\usage{
PCMPlotGaussianSample2D(mu, Sigma, numPoints = 1000, ...)
}
\arguments{
\item{mu}{numerical mean vector of length 2}

\item{Sigma}{numerical 2 x 2 covariance matrix}

\item{numPoints}{an integer denoting how many points should be randomly sampled (see details).}

\item{...}{additional arguments passed to ggplot.}
}
\value{
a ggplot object
}
\description{
A 2D sample from Gaussian distribution
}
\details{
This function generates a random sample of numPoints 2d points using
the function rmvnorm from the mvtnorm R-package. Then
it produces a ggplot on the generated points.
}
