% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMPlotTrajectory2D}
\alias{PCMPlotTrajectory2D}
\title{Plot the trajectory of the mean for a bivariate PCM over time starting from a value X0}
\usage{
PCMPlotTrajectory2D(X0s, model, regime = 1L, dims = c(1, 2),
  tX = seq(0, 100, by = 1), tVar = tX[seq(1, length(tX), length.out =
  4)], sizeSamp = 100, doPlot = TRUE, plot = NULL)
}
\description{
Plot the trajectory of the mean for a bivariate PCM over time starting from a value X0
}
\examples{
# a Brownian motion model with one regime
modelOU <- PCM(model = PCMDefaultModelTypes()['F'], k = 2)

# assign the model parameters at random: this will use uniform distribution
# with boundaries specified by PCMParamLowerLimit and PCMParamUpperLimit
# We do this in two steps:
# 1. First we generate a random vector. Note the length of the vector equals PCMParamCount(modelBM)
randomParams <- PCMParamRandomVecParams(modelOU, PCMNumTraits(modelOU), PCMNumRegimes(modelOU))
# 2. Then we load this random vector into the model.
PCMParamLoadOrStore(modelOU, randomParams, 0, PCMNumTraits(modelBM), PCMNumRegimes(modelBM), load = TRUE)
# let's plot the trajectory of the model starting from X0 = c(0,0)
PCMPlotTrajectory2D(c(0, 0), modelOU)
}
