% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMRandomVecParams}
\alias{PCMRandomVecParams}
\title{Generate a random parameter vector for a model using uniform distribution between its lower and upper bounds.}
\usage{
PCMRandomVecParams(model, n = 1L, argsPCMLowerBound = NULL,
  argsPCMUpperBound = NULL, argsPCMGetVecParams = NULL, ...)
}
\arguments{
\item{model}{a model PCM model object}

\item{n}{an integer specifying the number of random vectors to generate}

\item{argsPCMLowerBound, argsPCMUpperBound}{named lists of arguments passed to
\code{\link{PCMLowerBound}} and \code{\link{PCMUpperBound}}.}

\item{argsPCMGetVecParams}{a named list of argument passed to PCMGetVecPaarmas}

\item{...}{additional parameters.}
}
\value{
if n = 1, a numeric vector of length \code{PCMNumParams(model)}; if n > 1,
a numeric matrix of dimension n x \code{PCMNumParams(model)}.
}
\description{
Generate a random parameter vector for a model using uniform distribution between its lower and upper bounds.
}
\seealso{
PCMUpperBound PCMLowerBound PCMGetVecParams
}
