% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMSetOrGetVecParams.PCM}
\alias{PCMSetOrGetVecParams.PCM}
\title{Set or get the parameters of a BM model from or into a numeric vector}
\usage{
\method{PCMSetOrGetVecParams}{PCM}(model, vecParams, offset = 0, set = TRUE,
  ...)
}
\description{
Set or get the parameters of a BM model from or into a numeric vector
}
