% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSetOrGetVecParams}
\alias{PCMSetOrGetVecParams}
\title{Inplace set or get the parameters of a PCM from or into a numeric vector}
\usage{
PCMSetOrGetVecParams(model, vecParams, offset = 0, set = TRUE, ...)
}
\arguments{
\item{model}{a PCM model object}

\item{vecParams}{a numeric vector of length PCMNumParams(model) or longer}

\item{offset}{an integer indicating offset of the starting position for
writing/reading in vecParams. Default 0.}

\item{set}{a logical indicating whether to set the model parameters from
vecParams or to get them into vecParams. Default is TRUE which results in
setting the model parameters from vecParams. See details.}
}
\value{
an integer indicating the number of elements read/written from/to
vecParams.
}
\description{
Inplace set or get the parameters of a PCM from or into a numeric vector
}
\details{
This is an S3 generic. Note that depending on the value of \code{set},
this function changes either the model object or the vecParams vector from the
calling environment.
}
\seealso{
\code{\link{PCMGetVecParams}} \code{\link{PCMGetVecParamsFull}}
}
