% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSetRegimesAuto}
\alias{PCMSetRegimesAuto}
\title{Assign R regimes on a tree given a set of nodes}
\usage{
PCMSetRegimesAuto(tree, nodes)
}
\arguments{
\item{tree}{a phylo object}

\item{nodes}{an integer vector denoting tip or internal nodes in tree - the
regimes change at the start of the branches leading to these nodes.}
}
\value{
an integer vector of length nrow(tree$edge), with elements from
1 to (length(nodes) + 1) naming the regime for each edge in the tree. For
use in the PCM simulation and
likelihood calculation functions, this vector can be converted to character
and assigned as a member 'edge.regime' of the tree.
}
\description{
Assign R regimes on a tree given a set of nodes
}
