% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSim}
\alias{PCMSim}
\title{Simulation of a phylogenetic comparative model on a tree}
\usage{
PCMSim(tree, model, X0, metaI = PCMInfo(X = NULL, tree = tree, model = model,
  verbose = verbose), verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object specifying a rooted tree.}

\item{model}{an S3 object specifying the model (see Details).}

\item{X0}{a numeric vector of length k (the number of traits) specifying the
trait values at the root of the tree.}

\item{metaI}{a named list containg meta-information about the data and the
model.}

\item{verbose}{a logical indicating if informative messages should be written
during execution.}
}
\description{
Generate trait data on a tree according to a multivariate stochastic
model with one or several regimes
}
\details{
Internally, this function uses the \code{\link{PCMCond}} iimplementation
 for the given model class.
}
\seealso{
\code{\link{PCMLik}} \code{\link{PCMInfo}} \code{\link{PCMCond}}
}
