% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMSpecify}
\alias{PCMSpecify}
\title{Specify a phylogenetic comparative model for a given tree}
\usage{
PCMSpecify(tree, modelName, k, R, regimesUnique, paramNames, paramStorageModes,
  paramDims, ...)
}
\arguments{
\item{tree}{a phylo object}

\item{modelName}{a character specifying the name of the model. The
name of the model should be used as an S3 class of all model
instances.}

\item{k}{integer, specifying the number of traits;}

\item{R}{integer, specifying the number of regimes;}

\item{paramNames}{a list of characters specifying the names of
the model parameters. The length of the list is used to
determine the number of parameters.}

\item{paramStorageModes}{a list of characters of the same length
as `paramNames` specifying the storage.mode (native type) for
each parameter. Each element of this list should be one of
"double", "integer", "logical", "character", "complex", "raw".}

\item{paramDims}{a list of integer vectors, or character
strings evaluating as integer vectors specifying the
dimension of each model parameter for all model regimes.
This list should be the same length as `paramNames`. If an element of the
list is a character string it must evaluate to a vector of positive
integers, based on the objects accessible from the model parameters
(including k, R and ...) and
the symbols N and M, where N stays for the number of tips in the
tree, and M denotes the number of nodes. For example, if
there are 5 traits and 2 regimes, the dimension of the parameter Alpha of
an OU model should be specified as c(5, 5, 2) or "c(5, 5, 2)" or
"c(k, k, R)". If a parameter is a flat vector or an atom, for which the
dimension is NULL, specify an integer corresponding to its length, e.g.
the parameter specifying the presence of jumps in an OU model with jumps,
should have a dimension "M-1" corresponding to the number of edges in the
tree.}

\item{...}{additional arguments used in the evaluation of
`paramDims`.}
}
\value{
an S3 object of class ModelSpecification.
}
\description{
Specify a phylogenetic comparative model for a given tree
}
\seealso{
\code{\link{storage.mode}}
}
