% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{PCMSpecifyParams.BM}
\alias{PCMSpecifyParams.BM}
\title{Parameter specification of PCM model}
\usage{
\method{PCMSpecifyParams}{BM}(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a list
}
\description{
Parameter specification of PCM model
}
