% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMSpecifyParams}
\alias{PCMSpecifyParams}
\title{Parameter specification of PCM model}
\usage{
PCMSpecifyParams(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a list
}
\description{
Parameter specification of PCM model
}
