% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeDtNodeParts}
\alias{PCMTreeDtNodeParts}
\title{A data.table of the tips with their assigned part}
\usage{
PCMTreeDtNodeParts(tree)
}
\arguments{
\item{tree}{a phylo object with node-labels and parts}
}
\description{
A data.table of the tips with their assigned part
}
