% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeDtNodeRegimes}
\alias{PCMTreeDtNodeRegimes}
\title{A data.table of the tips with their assigned regime}
\usage{
PCMTreeDtNodeRegimes(tree)
}
\arguments{
\item{tree}{a phylo object with node-labels and regimes}
}
\description{
A data.table of the tips with their assigned regime
}
