% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeEdgeTimes}
\alias{PCMTreeEdgeTimes}
\title{A matrix with the begin and end time from the root for each edge in tree}
\usage{
PCMTreeEdgeTimes(tree)
}
\arguments{
\item{tree}{a phylo}
}
\description{
A matrix with the begin and end time from the root for each edge in tree
}
