% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeEvalNestedED}
\alias{PCMTreeEvalNestedED}
\title{Evaluate nested extractions (E) or deletions (D) of clades from a tree}
\usage{
PCMTreeEvalNestedED(expr)
}
\arguments{
\item{expr}{an R expression (or a character string evaluating to an R expression) of nested calls of functions
\code{E(x,node)} denoting extracting the clade rooted at node from the tree x, or \code{D(x,node)},
denoting dropping the clade rooted at node from the tree x. These calls can be nested, i.e. x can
be a symbol or r expression evaluating to a  phylo object in
the global or calling environment (corresponding to the original tree passed
 as argument) or a nested call to D or E.}
}
\value{
the resulting phylo object from evaluating expr in the calling environment
}
\description{
Evaluate nested extractions (E) or deletions (D) of clades from a tree
}
\seealso{
PCMTreeEvalNestedEDxOnTree
}
