% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeExtractBackboneRegimes}
\alias{PCMTreeExtractBackboneRegimes}
\title{Prune the tree leaving one tip for each regime}
\usage{
PCMTreeExtractBackboneRegimes(tree)
}
\arguments{
\item{tree}{a phylo with set node-labels and regimes}
}
\value{
a pruned version of tree
}
\description{
Prune the tree leaving one tip for each regime
}
\seealso{
PCMTreeSetLabels PCMTreeSetRegimes
}
