% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetBranchLength}
\alias{PCMTreeGetBranchLength}
\title{The length of the branch leading to a node}
\usage{
PCMTreeGetBranchLength(tree, daughterId)
}
\arguments{
\item{tree}{a phylo object.}

\item{daughterId}{an integer denoting the id of a daughter node}
}
\value{
a double denoting the length of the branch leading to daughterId
}
\description{
The length of the branch leading to a node
}
