% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetDaughters}
\alias{PCMTreeGetDaughters}
\title{A vector of the daughter nodes for a given parent node id in a tree}
\usage{
PCMTreeGetDaughters(tree, parentId)
}
\arguments{
\item{tree}{a phylo object.}

\item{parentId}{an integer denoting the id of the parent node}
}
\value{
an integer vector of the direct descendants of parentId
}
\description{
A vector of the daughter nodes for a given parent node id in a tree
}
