% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetLabels}
\alias{PCMTreeGetLabels}
\title{Get a vector of the tip and node labels in a tree}
\usage{
PCMTreeGetLabels(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
a character vector
}
\description{
Get a vector of the tip and node labels in a tree
}
