% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetParent}
\alias{PCMTreeGetParent}
\title{The parent node id of a daughter node in a tree}
\usage{
PCMTreeGetParent(tree, daughterId)
}
\arguments{
\item{tree}{a phylo object.}

\item{daughterId}{an integer denoting the id of the daughter node}
}
\value{
an integer denoting the parent of daughterId
}
\description{
The parent node id of a daughter node in a tree
}
