% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetPartNames}
\alias{PCMTreeGetPartNames}
\title{Unique parts on a tree in the order of occurrence from the root to the tips (preorder)}
\usage{
PCMTreeGetPartNames(tree)
}
\arguments{
\item{tree}{a phylo object with an additional member edge.part which should
be a character or an integer vector of length equal to the number of branches.}
}
\value{
a character vector.
}
\description{
Unique parts on a tree in the order of occurrence from the root to the tips (preorder)
}
