% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetPartRegimes}
\alias{PCMTreeGetPartRegimes}
\title{Regime mapping for the parts in a tree}
\usage{
PCMTreeGetPartRegimes(tree)
}
\arguments{
\item{tree}{a PCMTree or a phylo object.}
}
\value{
a named vector with names corresponding to the part names in tree
and values corresponding to regime names or ids.
}
\description{
Regime mapping for the parts in a tree
}
