% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetPartsForNodes}
\alias{PCMTreeGetPartsForNodes}
\title{Get the parts of the branches leading to a set of nodes or tips}
\usage{
PCMTreeGetPartsForNodes(tree, nodes)
}
\arguments{
\item{tree}{a phylo object with an edge.part member denoting parts.}

\item{nodes}{an integer vector denoting the nodes}
}
\value{
a character or an integer vector denoting the parts of the branches
leading to the nodes, according to tree$edge.part.
}
\description{
Get the parts of the branches leading to a set of nodes or tips
}
