% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetRegimeForNode}
\alias{PCMTreeGetRegimeForNode}
\title{Get the regime of the branch leading to a node or a tip}
\usage{
PCMTreeGetRegimeForNode(tree, node)
}
\arguments{
\item{tree}{a phylo object with an edge.regime member denoting regimes.}

\item{node}{an integer denoting the node}
}
\value{
a character or an integer denoting the regime of the branch leading
to the node, according to tree$edge.regime
}
\description{
Get the regime of the branch leading to a node or a tip
}
