% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetRegimesForNodes}
\alias{PCMTreeGetRegimesForNodes}
\title{Get the regimes of the branches leading to a set of nodes or tips}
\usage{
PCMTreeGetRegimesForNodes(tree, nodes)
}
\arguments{
\item{tree}{a phylo object with an edge.regime member denoting regimes.}

\item{node}{an integer vector denoting the nodes}
}
\value{
a character or an integer vector denoting the regimes of the branches
leading to the nodes, according to tree$edge.regime.
}
\description{
Get the regimes of the branches leading to a set of nodes or tips
}
