% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeJumps}
\alias{PCMTreeJumps}
\title{Jumps in modeled traits associated with branches in a tree}
\usage{
PCMTreeJumps(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
an integer vector of 0's and 1's with entries corresponding to the
denoting if a jump took place at the beginning of a branch.
}
\description{
Jumps in modeled traits associated with branches in a tree
}
