% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeListAllPartitions}
\alias{PCMTreeListAllPartitions}
\title{A list of all possible (including recursive) partitions of a tree}
\usage{
PCMTreeListAllPartitions(tree, minCladeSize, tableAncestors = NULL,
  verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object with set labels for the internal nodes}

\item{minCladeSize}{integer indicating the minimum number of tips allowed in
one part.}

\item{tableAncestors}{NULL (default) or an integer matrix returned by a
previous call to \code{PCMTreeTableAncestors(tree)}.}

\item{verbose}{a logical indicating if informative messages should be printed to
the console.}
}
\value{
a list of integer vectors.
}
\description{
A list of all possible (including recursive) partitions of a tree
}
