% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeLocateEpochOnBranches}
\alias{PCMTreeLocateEpochOnBranches}
\title{Find the crossing points of an epoch-time with each lineage of a tree}
\usage{
PCMTreeLocateEpochOnBranches(tree, epoch)
}
\arguments{
\item{tree}{a phylo}

\item{epoch}{a positive numeric indicating tip-ward distance from the root}
}
\value{
a named list with an integer vector element "nodes" denoting the ending nodes for each
branch crossing epoch and numeric vector element "positions" denoting the root-ward offset
from each node in nodes.
}
\description{
Find the crossing points of an epoch-time with each lineage of a tree
}
