% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeMatchRegimesWithModel}
\alias{PCMTreeMatchRegimesWithModel}
\title{Regimes associated with branches in a tree}
\usage{
PCMTreeMatchRegimesWithModel(tree, model)
}
\arguments{
\item{tree}{a phylo object}

\item{model}{a PCM object}
}
\value{
an integer vector with entries corresponding to the rows in tree$edge
  denoting the regime on each branch in the tree as an index in PCMRegimes(model).
}
\description{
Regimes associated with branches in a tree
}
