% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeMatrixNodesInSameRegime}
\alias{PCMTreeMatrixNodesInSameRegime}
\title{Which nodes in a tree belong to the same regime?}
\usage{
PCMTreeMatrixNodesInSameRegime(tree, upperTriangle = TRUE,
  returnVector = TRUE)
}
\arguments{
\item{tree}{a phylo object with an existing edge.regime}

\item{upperTriangle}{logical indicating if all duplicated entries and diagonal
entries sould be set to NA (by default TRUE).}

\item{returnVector}{logical indicating if a vector instead of a matrix should be
returned (corresponding to calling as.vector on the resulting matrix and removing
NAs). Default: TRUE}
}
\value{
a M x M logical matrix with TRUE on the diagonal and for each couple of
tips that belong to the same dagonal. if returnVector is TRUE (default) only a
vector of the non-NA entries will be returned.
}
\description{
Which nodes in a tree belong to the same regime?
}
