% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeNearestNodesToEpoch}
\alias{PCMTreeNearestNodesToEpoch}
\title{Find the nearest node to a given time from the root (epoch) on each lineage crossing this epoch}
\usage{
PCMTreeNearestNodesToEpoch(tree, epoch)
}
\arguments{
\item{tree}{a phylo}

\item{epoch}{a positive numeric}
}
\value{
an integer vector
}
\description{
Find the nearest node to a given time from the root (epoch) on each lineage crossing this epoch
}
