% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeNumParts}
\alias{PCMTreeNumParts}
\title{Number of unique parts on a tree}
\usage{
PCMTreeNumParts(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
the number of different parts encountered on the tree branches
}
\description{
Number of unique parts on a tree
}
