% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMTreeNumTips}
\alias{PCMTreeNumTips}
\title{Wrapper for length(tree$tip.label)}
\usage{
PCMTreeNumTips(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
the number of tips in tree
}
\description{
Wrapper for length(tree$tip.label)
}
