% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMTreeNumUniqueRegimes}
\alias{PCMTreeNumUniqueRegimes}
\title{Number of unique regimes on a tree}
\usage{
PCMTreeNumUniqueRegimes(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
the number of different regimes encountered on the tree branches
}
\description{
Number of unique regimes on a tree
}
