% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreePlot}
\alias{PCMTreePlot}
\title{Plot a tree with parts}
\usage{
PCMTreePlot(tree, palette = PCMColorPalette(PCMTreeNumParts(tree),
  PCMTreeGetPartNames(tree)), ...)
}
\arguments{
\item{tree}{a phylo with set labels and parts}

\item{palette}{a named vector of colors corresponding to the parts in tree}

\item{...}{Arguments passed to ggtree, e.g. layout = 'fan', open.angle = 8, size=.25.}
}
\description{
Plot a tree with parts
}
\note{
Currently, the ggtree package is not on CRAN and therefore it is not explicitly
imported by PCMBase.
}
