% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreePostorder}
\alias{PCMTreePostorder}
\title{Post-order tree traversal}
\usage{
PCMTreePostorder(tree)
}
\arguments{
\item{tree}{a phylo object with possible singleton nodes (i.e. internal nodes
with one daughter node)}
}
\value{
a vector of indices of edges in tree$edge in post-order.
}
\description{
Post-order tree traversal
}
