% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeSetDefaultPartition}
\alias{PCMTreeSetDefaultPartition}
\title{Set a default partition in a tree}
\usage{
PCMTreeSetDefaultPartition(tree, part)
}
\arguments{
\item{tree}{a phylo object}

\item{part}{a character, an integer or PCM model object}
}
\value{
This function does not return a value but has a side effect on the
 passed tree object.
}
\description{
This function sets or overwrites the current member edge.part
in tree with one of the following:
\itemize{
\item{rep(part[1], length(tree$edge.length))}{if part is a character or an integer}
\item{rep(PCMRegimes(part)[1], length(tree$edge.length))}{if part is a PCM model}
} Note that the function modifies the passed tree object inplace.
}
