% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMTreeSetRegimesIncremental}
\alias{PCMTreeSetRegimesIncremental}
\title{Assign regimes on a tree given a set of starting branches}
\usage{
PCMTreeSetRegimesIncremental(tree, nodes)
}
\arguments{
\item{tree}{a phylo object}

\item{nodes}{an integer vector denoting tip or internal nodes in tree - the
regimes change at the start of the branches leading to these nodes.}
}
\value{
an integer vector of length nrow(tree$edge), with elements from
1 to (length(nodes) + 1) naming the regime for each edge in the tree. For
use in the PCM simulation and
likelihood calculation functions, this vector can be converted or mapped to a
a vector of character strings and/or assigned as a member 'edge.regime' of the tree.
}
\description{
Assign regimes on a tree given a set of starting branches
}
\details{
It is assumed that each regime "paints" a linked subset of branches
on a tree. Thus, each regime is fully described by its starting branch. The
descendant branches inherit this regime until reaching a tip or a node that is
present in the \code{nodes} parameter.
}
