% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeToString}
\alias{PCMTreeToString}
\title{A character representation of a phylo object.}
\usage{
PCMTreeToString(tree, includeLengths = FALSE, includePartition = FALSE)
}
\arguments{
\item{tree}{a phylo object.}

\item{includeLengths}{logical. Default: FALSE.}

\item{includePartition}{logical. Default: FALSE.}
}
\value{
a character string.
}
\description{
A character representation of a phylo object.
}
