% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeVCV}
\alias{PCMTreeVCV}
\title{Phylogenetic Variance-covariance matrix}
\usage{
PCMTreeVCV(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
a N x N matrix. Assuming a BM model of evolution, this is a matrix
in which element (i,j) is equal to the shared root-distance of the nodes i
and j.
}
\description{
This is a simplified wrapper for ape's \code{\link{vcv}} function. Setting
the runtime option PCMBase.UsePCMVarForVCV to TRUE will switch to a
computation of the matrix using the function \code{\link{PCMVar}}.
}
\seealso{
\code{\link{vcv}} \code{\link{PCMVar}} \code{\link{PCMOptions}}
}
