% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMUnfixParameter}
\alias{PCMUnfixParameter}
\title{Unfix a parameter in a PCM model}
\usage{
PCMUnfixParameter(model, name)
}
\arguments{
\item{model}{a PCM object}

\item{name}{a character string}
}
\value{
a copy of the model with removed class '_Fixed' from the class of the
parameter \code{name}
}
\description{
Unfix a parameter in a PCM model
}
