% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMUniqueRegimesTree}
\alias{PCMUniqueRegimesTree}
\title{Unique regimes on a tree}
\usage{
PCMUniqueRegimesTree(tree)
}
\arguments{
\item{tree}{a phylo object with an additional member edge.regime which should
be a character or an integer vector of length equal to the number of branches.}
}
\value{
a character or an integer vector depending on tree$edge.regime.
}
\description{
This is a shortcut for sort(unique(tree$edge.regime))
}
