% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMUpperBound}
\alias{PCMUpperBound}
\title{Numerical upper bound}
\usage{
PCMUpperBound(model, upperBoundValue = 10, upperBoundValuePositiveDiag = 10,
  namedUpperBoundValues = NULL, ...)
}
\arguments{
\item{model}{a PC
M object}
}
\value{
a PCM object of the same S3 classes as model. Calling
\code{\link{PCMSetOrGetVecParams}} on this object returns an upper
bound for that can be used, e.g. in a call to \code{\link{optim}}
#' model <- PCM("BM__NoX0__NoSigmae_x", k = 3)
PCMLowerBound(model)
}
\description{
Numerical upper bound
}
