% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{PCMValidate.TwoSpeedOU}
\alias{PCMValidate.TwoSpeedOU}
\title{Validate TwoSpeedOU parameters}
\usage{
\method{PCMValidate}{TwoSpeedOU}(tree, model, verbose = FALSE)
}
\description{
Validate TwoSpeedOU parameters
}
