% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMValidateGeneral}
\alias{PCMValidateGeneral}
\title{General function for validating a model}
\usage{
PCMValidateGeneral(tree, model, modelSpec, verbose)
}
\arguments{
\item{tree}{an object of S3 class "phylo"}

\item{model}{a model object of S3 class the name of the model}

\item{modelSpec}{an object of S3 class ModelSpecification}

\item{verbose}{a logical indicating if the log messges should be
printed on the screen during validation.}
}
\value{
a named list
}
\description{
General function for validating a model
}
