% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMVar}
\alias{PCMVar}
\title{Expected variance-covariance matrix for each couple of tips (i,j)}
\usage{
PCMVar(tree, model, W0 = matrix(0, PCMNumTraits(model),
  PCMNumTraits(model)), metaI = PCMInfo(NULL, tree, model, verbose),
  internal = FALSE, verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{W0}{a numeric matrix denoting the initial k x k variance covariance matrix at the
root (default is the k x k zero matrix).}

\item{metaI}{a list returned from a call to \code{PCMInfo(X, tree, model)},
containing meta-data such as N, M and k.}

\item{internal}{a logical indicating if the per-node variance-covariances matrices for
the internal nodes should be returned (see Value). Default FALSE.}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\value{
If internal is FALSE, a (k x N) x (k x N) matrix W, such that k x k block
\code{W[((i-1)*k)+(1:k), ((j-1)*k)+(1:k)]} equals the expected
covariance matrix between tips i and j. Otherwise, a list with an element 'W' as described above and
a k x M matrix element 'Wii' containing the per-node variance covariance matrix for each node:
The k x k block \code{Wii[, (i-1)*k + (1:k)]} represents the variance covariance matrix for node i.
}
\description{
Expected variance-covariance matrix for each couple of tips (i,j)
}
