% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMVarAtTime}
\alias{PCMVarAtTime}
\title{Calculate the variance covariance k x k matrix at time t, under a PCM model}
\usage{
PCMVarAtTime(t, model, W0 = matrix(0, PCMNumTraits(model),
  PCMNumTraits(model)), regime = 1L, verbose = FALSE)
}
\arguments{
\item{t}{positive numeric denoting time}

\item{model}{a PCM model object}

\item{W0}{a numeric matrix denoting the initial k x k variance covariance matrix at the
root (default is the k x k zero matrix).}

\item{regime}{an integer or a character denoting the regime in model for which to do the calculation;
(Defaults to 1L meaning the first regime in the model)}

\item{verbose}{a logical indicating if (debug) messages should be written on the console (Defaults to FALSE).}
}
\value{
A numeric k x k matrix
}
\description{
Calculate the variance covariance k x k matrix at time t, under a PCM model
}
