% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{presentCoordinates}
\alias{presentCoordinates}
\title{For every node (root, internal or tip) in tree, build a logical vector of
length k with TRUE values for every present coordinate.}
\usage{
presentCoordinates(X, tree, pruneI = pruneTree(tree))
}
\arguments{
\item{X}{numeric k x N matrix of observed values, with possible NA entries. The
columns in X are in the order of tree$tip.label}

\item{tree}{a phylo object}

\item{pruneI}{a list returned by the pruneTree function. Either leave this
as default or pass a previously computed pruneI for the same tree.}
}
\value{
a k x M logical matrix
}
\description{
For every node (root, internal or tip) in tree, build a logical vector of
length k with TRUE values for every present coordinate.
}
