% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{pruneTree}
\alias{pruneTree}
\title{Extract information for pruning a tree used as cache in poumm likelihood
calculation}
\usage{
pruneTree(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
a list of objects
}
\description{
Extract information for pruning a tree used as cache in poumm likelihood
calculation
}
\details{
This method should only be called if calculating poumm likelihood
with impl='R5'.
}

