% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{RequireSuggestedPackages}
\alias{RequireSuggestedPackages}
\title{Check if all packages listed in Suggests are available}
\usage{
RequireSuggestedPackages()
}
\value{
logical TRUE if suggested packages are installed and can be loaded; FALSE
otherwise
}
\description{
Check if all packages listed in Suggests are available
}
