% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{texp.JOU}
\alias{texp.JOU}
\title{Create a function of time that calculates exp(-H*time)}
\usage{
texp.JOU(H)
}
\arguments{
\item{H}{a kxk matrix specifying the alphas in a JOU process}
}
\value{
a function of time returning a matrix
}
\description{
Create a function of time that calculates exp(-H*time)
}
