% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{V.BM}
\alias{V.BM}
\title{Generate a multivariate (MV) BM variance-covariance function}
\usage{
V.BM(Sigma, threshold0 = 0)
}
\arguments{
\item{Sigma}{the matrix Sigma of a MV BM process.}
}
\value{
a function of one numerical argument (time), which calculates the
expected variance covariance matrix of a MV-BM process after time, given
the specified arguments.
}
\description{
Generate a multivariate (MV) BM variance-covariance function
}
