% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{V.JOU}
\alias{V.JOU}
\title{Generate a multivariate (MV) JOU variance-covariance function}
\usage{
V.JOU(lambda, P, P_1, Sigma, Alpha, Sigmaj, threshold0 = 0)
}
\arguments{
\item{lambda}{vector of the eignevalues of the matrix Alpha of a MV JOU process.}

\item{P}{matrix of the eigenvectors of the matrix Alpha}

\item{P_1}{inverse eigenvectors matrix}

\item{Sigma}{the matrix Sigma of a MV JOU process.}

\item{Alpha}{is the matrix Alpha of a MV JOU process.}

\item{Sigmaj}{is the variance matrix of the jump distribution in JOU process}
}
\value{
a function of two numerical arguments (time) and (xi:binary value denoting jump or not) , which calculates the
expected variance covariance matrix of a MV-JOU process after time, given
the specified arguments.
}
\description{
Generate a multivariate (MV) JOU variance-covariance function
}
