% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{V.TwoSpeedOU}
\alias{V.TwoSpeedOU}
\title{Generate a multivariate (MV) TwoSpeedOU variance-covariance function}
\usage{
V.TwoSpeedOU(lambda, P, P_1, Sigma, threshold0 = 0)
}
\arguments{
\item{lambda}{vector of the eignevalues of the matrix Alpha2 of a MV TwoSpeedOU process.}

\item{P}{matrix of the eigenvectors of the matrix Alpha}

\item{P_1}{inverse eigenvectors matrix}

\item{Sigma}{the matrix Sigma of a MV TwoSpeedOU process.}
}
\value{
a function of one numerical argument (time), which calculates the
expected variance covariance matrix of a MV-TwoSpeedOU process after time, given
the specified arguments.
}
\description{
Generate a multivariate (MV) TwoSpeedOU variance-covariance function
}
