% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{validateModel}
\alias{validateModel}
\title{An S3 generic function validating a model for a given tree.}
\usage{
validateModel(tree, model, verbose = FALSE)
}
\description{
This function should be implemented for each model.
}
