% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{validateModel.TwoSpeedOU}
\alias{validateModel.TwoSpeedOU}
\title{Validate TwoSpeedOU parameters}
\usage{
\method{validateModel}{TwoSpeedOU}(tree, model, verbose = FALSE)
}
\description{
Validate TwoSpeedOU parameters
}
