% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/White.R
\name{White}
\alias{White}
\title{White Gaussian PCM ignoring phylogenetic history}
\description{
White model ignoring phylogenetic history, treating trait values
 as independent samples from a k-variate Gaussian.
}
\details{
Calculating likelihoods for this model does not work if the global
option PCMBase.Singular.Skip is set to FALSE.
}
