% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{AbCdEf}
\alias{AbCdEf}
\title{Quadratic polynomial parameters A, b, C, d, E, f for each node}
\usage{
AbCdEf(tree, model, metaI, pc, verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{metaI}{a list returned from a call to \code{validateModel(tree, model)},
containing meta-data such as N, M and k.}

\item{pc}{a logical k x M matrix returned by \code{\link{presentCoordinates}}.}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\description{
An S3 generic function that has to be implemented for every
 model class. This function is called by \code{\link{mvlik}}.
}
