% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{Args_MixedGaussian_MGPMDefaultModelTypes}
\alias{Args_MixedGaussian_MGPMDefaultModelTypes}
\title{Arguments to be passed to the constructor MixedGaussian when constructing
a MGPM model with some of the default MGPM model types.}
\usage{
Args_MixedGaussian_MGPMDefaultModelTypes()
}
\value{
a list of named arguments.
}
\description{
Arguments to be passed to the constructor MixedGaussian when constructing
a MGPM model with some of the default MGPM model types.
}
\seealso{
MGPMDefaultModelTypes
}
