% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{as.MGPM}
\alias{as.MGPM}
\title{If necessary, convert an object to an MGPM object.}
\usage{
as.MGPM(s, ctx = NULL)
}
\arguments{
\item{s}{a \code{\link{MGPM}} or a \code{\link{MGPMVector}}
object.}

\item{ctx}{a \code{\link{MGPMContext}} object. This is needed only if
\code{s} is \code{\link{MGPMVector}} object.}
}
\value{
an MGPM object corresponding to \code{s}. If \code{s} is already
a \code{\link{MGPM}} object, it is returned as is. Otherwise, a
\code{\link{MGPM}} object is constructed from \code{s} and \code{ctx}.
}
\description{
If necessary, convert an object to an MGPM object.
}
