% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU.R
\name{fLambda_ij.OU}
\alias{fLambda_ij.OU}
\title{Create a function of time that calculates (1-exp(-lambda_{ij}*time))/lambda_{ij}
for every element lambda_{ij} of the input matrix Lambda_ij}
\usage{
fLambda_ij.OU(Lambda_ij, threshold0 = 0)
}
\arguments{
\item{Lambda_ij}{a squared numerical matrix of dimension k x k}

\item{threshold0}{numeric threshold value. For |Lambda_{ij}| smaller or equal
than threshold the limit time is returned.}
}
\value{
a function of time returning a matrix with entries formed from the
 above function or the limit time if |Lambda_{ij}|<=trehshold0.
}
\description{
Create a function of time that calculates (1-exp(-lambda_{ij}*time))/lambda_{ij}
for every element lambda_{ij} of the input matrix Lambda_ij
}
