% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatexTools.R
\name{FormatCellAsLatex}
\alias{FormatCellAsLatex}
\title{Latex represenataion of a model parameter or other found in a data.table object}
\usage{
FormatCellAsLatex(x)
}
\arguments{
\item{x}{an R object. Currently, character vectors of length 1,
numeric vectors and matrices are supported.}
}
\value{
a character string
}
\description{
Latex represenataion of a model parameter or other found in a data.table object
}
