% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{MGPMContext}
\alias{MGPMContext}
\title{Context of a MGPM}
\usage{
MGPMContext(X, tree, model, SE = matrix(0, PCMNumTraits(model),
  PCMTreeNumTips(tree)), K = 0L, n = integer(0L), l = double(0L),
  r = N + 1L, m = 1L, v = NULL)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} and \code{NaN} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Missing values can be either not-available (\code{NA}) or not existing (\code{NaN}).
These two values have are treated differently when calculating
likelihoods: see \code{\link{PCMPresentCoordinates}}.}

\item{tree}{a phylo object with N tips.}

\item{model}{a template used to build MixedGaussian model objects.}

\item{SE}{a k x N matrix specifying the standard error for each measurement in
X. Alternatively, a k x k x N cube specifying an upper triangular k x k
factor of the variance covariance matrix for the measurement error
for each node i=1, ..., N.
Default: \code{matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree))}.}

\item{K, n, l, r, m, v}{default (or initial) variables for the number of
shifts, the shift nodes, the shift-offsets, the regimes, the mapped model
types and the parameter values (see \code{\link{MGPM}}).
By default, these are
\code{K = 0L, n = integer(0), l = double(0), r = N+1L, m = 1L, v = double(P)}.
Here, N is the number of tips in the tree, and P is the number of numerical
parameters in a MixedGaussian with a single regime mapped to model type 1.}
}
\value{
an object of S3 class 'MGPMContext'.
}
\description{
Context of a MGPM
}
