% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRG.R
\name{MRG}
\alias{MRG}
\alias{is.MRG}
\alias{MRG}
\title{Multiple regime Gaussian PCMs}
\usage{
is.MRG(x)

MRG(k, models, mapping, className = paste0("MRG_", do.call(paste0,
  as.list(mapping))), X0 = list(default = rep(0, k), type = c("gvector",
  "full"), description = "trait vector at the root; global for all regimes"),
  Sigmae_x = list(default = matrix(0, nrow = k, ncol = k), type = c("gmatrix",
  "upper.tri.diag", "positive.diag"), description =
  "Upper triangular Choleski factor of the variance-covariance matrix for the non-phylogenetic trait component"))
}
\arguments{
\item{k}{integer defining the number of traits.}

\item{models}{a character string vector with the class names of the possible
models included in the MRG, e.g. c("BM3", "OU3").}

\item{mapping}{a character string vector with elements from models or an
integer vector with elements between 1 and length(models)
mapping models to regimes (the positions in the mapping vector), e.g.
c(a = 1, b = 1, c = 2, d = 1) defines an MRG with four different regimes with
models BM3, BM3, OU3 and BM3, corresponding to each regime.}

\item{className}{a character string definingn a valid class name for the
resulting MRG object.}

\item{X0}{NULL or a list defining a global vector X0 to be used by all
models in the MRG.}

\item{Sigmae_x}{NULL or a list defining a global matrix to be used as element
Sigmae_x by all models in the MRG.}
}
\value{
an object of S3 class className inheriting from MRG, GaussianPCM and
PCM.
}
\description{
Create a multi-regime Gaussian model (MRG)
}
\details{
If X0 is not NULL it has no sense to use models including X0 as a
parameter (e.g. use BM1 or BM3 insted of BM or BM2). Similarly if Sigmae_x is
not NULL there is no meaning in using models including Sigmae_x as a parameter,
(e.g. use OU2 or OU3 instead of OU or OU1).
}
