% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{mvcond}
\alias{mvcond}
\title{Multivariate normal distribution for a given tree and model}
\usage{
mvcond(tree, model, metaI, r = 1, verbose = FALSE)
}
\description{
Multivariate normal distribution for a given tree and model
}
