% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCM.PCM}
\alias{PCM.PCM}
\title{Create a PCM}
\usage{
\method{PCM}{PCM}(model, k = 1, regimes = 1, params = NULL,
  vecParams = NULL, offset = 0, ...)
}
\arguments{
\item{model}{This argument can take one of the following forms:
\item a character vector of the S3-classes of the model object to be
created (one model object can have one or more S3-classes);
\item a list of character vectors as above denoting the S3 classes of each
regime in a generalized multiple regime model. If available, the names of the
list-members will be used as regime names;
\item an S3 object representing a PCM model}
}
\description{
Create a PCM
}
