% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMBaseIsADevRelease}
\alias{PCMBaseIsADevRelease}
\title{Check if the PCMBase version correpsonds to a dev release}
\usage{
PCMBaseIsADevRelease()
}
\value{
a logical
}
\description{
We define a dev release as having a sub-release, eg 0.9.15.5 is
one whereas 0.9.16 is not.
}
