% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{PCMCond.TwoSpeedOU}
\alias{PCMCond.TwoSpeedOU}
\title{Conditional multivariate TwoSpeedOU distribution}
\usage{
\method{PCMCond}{TwoSpeedOU}(tree, model, r = 1, metaI = PCMInfo(NULL, tree,
  model, verbose), verbose = FALSE)
}
\arguments{
\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{r}{an integer specifying a model regime}

\item{metaI}{a list returned from a call to \code{PCMValidate(tree, model)},
containing meta-data such as N, M and k.}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\value{
a named list as specified in \code{\link{PCMCond}}
}
\description{
Conditional multivariate TwoSpeedOU distribution
}
