% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultModelTypes.R
\name{PCMDefaultModelTypes}
\alias{PCMDefaultModelTypes}
\alias{MGPMDefaultModelTypes}
\title{Class names for the the default PCM and MGPM model types}
\usage{
PCMDefaultModelTypes()

MGPMDefaultModelTypes()
}
\value{
a character string vector with named elements
(currently  A,B,C,D,E,F).
}
\description{
Utility functions returning named character vector of the
model class-names for the default model types used for PCM and MixedGaussian
model construction.
}
\seealso{
Args_MixedGaussian_MGPMDefaultModelTypes
}
