% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMInfo}
\alias{PCMInfo}
\title{Meta-information about a tree associated with a PCM}
\usage{
PCMInfo(X, tree, model, verbose = FALSE)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Depending on the value of the global option "PCMBase.Internal.PC.Full" \code{NA}
entries are interpreted either as missing measurements or non-existing traits
(see \code{\link{PCMPresentCoordinates}}).}

\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\value{
a named list with the following elements:
\item{M}{total number of nodes in the tree;}
\item{N}{number of tips;}
\item{k}{number of traits;}
\item{RTree}{number of regimes on the tree (distinct elements of tree$edge.regime);}
\item{RModel}{number of regimes in the model (distinct elements of attr(model, regimes));}
\item{p}{number of free parameters describing the model;}
\item{r}{an integer vector corresponding to tree$edge with the regime for each
branch in tree;}
\item{xi}{an integer vector of 0's and 1's corresponding to the rows in tree$edge
indicating the presence of a jump at the corresponding branch;}
\item{pc}{a logical matrix of dimension k x M denoting the present coordinates
for each node;}
}
