% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMListParameterizations}
\alias{PCMListParameterizations}
\title{Specify the parameterizations for each parameter of a model}
\usage{
PCMListParameterizations(model, ...)
}
\description{
Specify the parameterizations for each parameter of a model
}
