% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMMeanAtTime}
\alias{PCMMeanAtTime}
\title{Calculate the mean at time t, given X0, under a PCM model}
\usage{
PCMMeanAtTime(t, model, X0 = model$X0, regime = 1L, verbose = FALSE)
}
\arguments{
\item{t}{positive numeric denoting time}

\item{model}{a PCM model object}

\item{X0}{a numeric vector of length k, where k is the number of traits in the model (Defaults to model$X0).}

\item{regime}{an integer or a character denoting the regime in model for which to do the calculation;
(Defaults to 1L meaning the first regime in the model)}

\item{verbose}{a logical indicating if (debug) messages should be written on the console (Defaults to FALSE).}
}
\value{
A numeric vector of length k
}
\description{
Calculate the mean at time t, given X0, under a PCM model
}
