% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{PCMNumRegimes}
\alias{PCMNumRegimes}
\title{Number of regimes on a tree modeled with PCM}
\usage{
PCMNumRegimes(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
the number of different regimes encountered on the tree branches
}
\description{
Number of regimes on a tree modeled with PCM
}
