% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMParamLoadOrStore}
\alias{PCMParamLoadOrStore}
\title{Load a PCM parameter from a vector of the variable parameters in a model.}
\usage{
PCMParamLoadOrStore(o, vecParams, offset, k, R, load, parentModel = NULL)
}
\value{
an integer equaling the number of elemnents read from vecParams.
In the case of type=="custom", the number of indices bigger than offset returned by the function indices(offset, k).
}
\description{
Load a PCM parameter from a vector of the variable parameters in a model.
}
\details{
This S3 generic function has both, a returned value and side effects.
}
