% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMParamLowerLimit}
\alias{PCMParamLowerLimit}
\title{The object representing a lower limit for a given type}
\usage{
PCMParamLowerLimit(o, k, R, ...)
}
\arguments{
\item{o}{an object such as a VectorParameter a MatrixParameter or a PCM.}

\item{k}{integer denoting the number of traits}

\item{R}{integer denoting the number of regimes in the model in which o
belongs to.}

\item{...}{additional arguments (optional or future use).}
}
\value{
an object of the same S3 class as o representing a lower limit for
the class.
}
\description{
\code{PCMParamLowerLimit} and \code{PCMParamUpperLimit} are S3
generic functions.
}
