% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMParam.R
\name{PCMParamSetByName}
\alias{PCMParamSetByName}
\title{Set model parameters from a named list}
\usage{
PCMParamSetByName(model, params, inplace = TRUE,
  replaceWholeParameters = FALSE, ...)
}
\arguments{
\item{model}{a PCM model object}

\item{params}{a named list with elements among the names found in model}

\item{inplace}{logical indicating if the parameters should be set "inplace" for the
model object in the calling environment or a new model object with the parameters set
as specified should be returned. Defaults to TRUE}

\item{tree}{a phylo object (possible future use)}
}
\value{
If inplace is TRUE, the function only has a side effect of setting the
parameters of the model object in the calling environment; otherwise the function
returns a modified copy of the model object.
}
\description{
Set model parameters from a named list
}
