% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMPlotMath}
\alias{PCMPlotMath}
\title{Beautiful model description based on plotmath}
\usage{
PCMPlotMath(o, roundDigits = 2, transformSigma_x = FALSE)
}
\arguments{
\item{o}{a PCM or a parameter object.}

\item{roundDigits}{an integer, default: 2.}

\item{transformSigma_x}{a logical indicating if Cholesky transformation should be
applied to Cholesky-factor parameters prior to generating the plotmath expression.}
}
\value{
a character string.
}
\description{
This is an S3 generic that produces a plotmath expression for
its argument.
}
