% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMPlotTraitData2D}
\alias{PCMPlotTraitData2D}
\title{Scatter plot of 2-dimensional data}
\usage{
PCMPlotTraitData2D(X, tree, legend = FALSE, labeledTips = NULL)
}
\description{
Scatter plot of 2-dimensional data
}
