% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMRegimes}
\alias{PCMRegimes}
\title{Get the regimes (aka colors) of a PCM or of a PCMTree object}
\usage{
PCMRegimes(obj)
}
\arguments{
\item{obj}{a PCM or a PCMTree object}
}
\value{
a character or an integer vector giving the regime names in the obj
}
\description{
Get the regimes (aka colors) of a PCM or of a PCMTree object
}
