% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMRegimes}
\alias{PCMRegimes}
\title{Regimes in a model}
\usage{
PCMRegimes(model)
}
\arguments{
\item{model}{a PCM object}
}
\value{
a character or an integer vector giving the regime names of the models
}
\description{
Regimes in a model
}
