% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSetOrGetVecParams}
\alias{PCMSetOrGetVecParams}
\title{Inplace set or get the parameters of a PCM from or into a numeric vector}
\usage{
PCMSetOrGetVecParams(model, vecParams, offset = 0, set = TRUE, ...)
}
\description{
Inplace set or get the parameters of a PCM from or into a numeric vector
}
