% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSpecify}
\alias{PCMSpecify}
\title{Parameter specification of PCM model}
\usage{
PCMSpecify(model, ...)
}
\arguments{
\item{model}{a PCM model object}
}
\value{
a list
}
\description{
Parameter specification of PCM model
}
