% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM.R
\name{PCMSpecify}
\alias{PCMSpecify}
\title{Parameter specification of PCM model}
\usage{
PCMSpecify(model, ...)
}
\arguments{
\item{model}{a PCM model object.}

\item{...}{additional arguments used by implementing methods.}
}
\value{
a list specifying the parameters of a PCM.
}
\description{
The parameter specification of a PCM model represents a named
list with an entry for each parameter of the model. Each entry in the list is
a structure defining the S3 class of the parameter and its verbal description.
This is an S3 generic. See `PCMSpecify.OU` for an example method.
}
