% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTable.R
\name{PCMTable}
\alias{PCMTable}
\title{A data.table representation of a PCM object}
\usage{
PCMTable(model, skipGlobalRegime = FALSE)
}
\arguments{
\item{model}{a PCM object.}

\item{skipGlobalRegime}{logical indicating whether a raw in the returned
table for the global-scope parameters should be omitted (this is mostly for
internal use). Default (FALSE).}
}
\value{
an object of S3 class PCMTable
}
\description{
A data.table representation of a PCM object
}
\details{
This is an S3 generic.
}
