% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeBackbonePartition}
\alias{PCMTreeBackbonePartition}
\title{Prune the tree leaving one tip for each part}
\usage{
PCMTreeBackbonePartition(tree)
}
\arguments{
\item{tree}{a phylo with set node-labels and parts}
}
\value{
a pruned version of tree
}
\description{
Prune the tree leaving one tip for each part
}
\seealso{
PCMTreeSetLabels PCMTreeSetPartition
}
