% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetLabels}
\alias{PCMTreeGetLabels}
\alias{PCMTreeGetRootLabel}
\alias{PCMTreeGetNodeLabels}
\alias{PCMTreeGetTipLabels}
\title{Node labels of a tree}
\usage{
PCMTreeGetLabels(tree)

PCMTreeGetRootLabel(tree)

PCMTreeGetNodeLabels(tree)

PCMTreeGetTipLabels(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
a character vector
}
\description{
Get the character labels of the tips, root and internal nodes in the tree
(see Functions below).
}
\section{Functions}{
\itemize{
\item \code{PCMTreeGetLabels()}: Get all labels in the order (tips,root,internal).

\item \code{PCMTreeGetRootLabel()}: Get the root label

\item \code{PCMTreeGetNodeLabels()}: Get the internal node labels

\item \code{PCMTreeGetTipLabels()}: Get the tip labels

}}
