% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetRegimesForEdges}
\alias{PCMTreeGetRegimesForEdges}
\title{Model regimes (i.e. colors) associated with the branches in a tree}
\usage{
PCMTreeGetRegimesForEdges(tree)
}
\arguments{
\item{tree}{a PCMTree or a phylo object.}
}
\value{
a vector with entries corresponding to the rows in tree$edge
  denoting the regime associated with each branch in the tree. The type
  of the vector element is defined by the type of tree$part.regime.
}
\description{
Model regimes (i.e. colors) associated with the branches in a tree
}
