% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetRegimesForNodes}
\alias{PCMTreeGetRegimesForNodes}
\title{Get the regimes of the branches leading to a set of nodes or tips}
\usage{
PCMTreeGetRegimesForNodes(tree, nodes = seq_len(PCMTreeNumNodes(tree)))
}
\arguments{
\item{tree}{a phylo object with an edge.part member denoting parts.}

\item{nodes}{an integer vector denoting the nodes.
Default is seq_len(PCMTreeNumNodes(tree).}
}
\value{
a character vector denoting the parts of the branches
leading to the nodes, according to tree$edge.part.
}
\description{
Get the regimes of the branches leading to a set of nodes or tips
}
