% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeGetTipsInRegime}
\alias{PCMTreeGetTipsInRegime}
\title{Get the tips belonging to a regime tree}
\usage{
PCMTreeGetTipsInRegime(tree, regime)
}
\arguments{
\item{tree}{a phylo object with an edge.regime member}

\item{regime}{a character or integer belonging to tree$edge.regime}
}
\value{
an integer vector with the ids of the tips belonging to regime
}
\description{
Get the tips belonging to a regime tree
}
