% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMTree.R
\name{PCMTreeMatchLabels}
\alias{PCMTreeMatchLabels}
\title{Get the node numbers associated with tip- or node-labels in a tree}
\usage{
PCMTreeMatchLabels(tree, labels)
}
\arguments{
\item{tree}{a phylo object}

\item{labels}{a character vector with valid tip or node labels from tree}
}
\value{
an integer vector giving the tip- or node- integer indices
 corresponding to labels.
}
\description{
Get the node numbers associated with tip- or node-labels in a tree
}
