% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariatePCM.R
\name{pruneTree}
\alias{pruneTree}
\title{Extract information for pruning a tree used as cache during likelihood
calculation}
\usage{
pruneTree(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
a list of objects used by the R implementation of PCMLik().
}
\description{
Extract information for pruning a tree used as cache during likelihood
calculation
}
\details{
The function is very slow on strongly unbalanced trees due to the
slow vector append() operation in R.
}
