% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{texp.JOU}
\alias{texp.JOU}
\title{Create a function of time that calculates exp(-Alpha*time)}
\usage{
texp.JOU(Alpha, threshold0 = 0)
}
\arguments{
\item{threshold0}{numeric threshold value.}

\item{the}{argument Alpha is an Alpha matrix kxk specifying the alphas in a JOU process}
}
\value{
a function of time returning a matrix
}
\description{
Create a function of time that calculates exp(-Alpha*time)
}
