% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOU.R
\name{V.OU}
\alias{V.OU}
\title{Generate a multivariate (MV) OU variance-covariance function}
\usage{
V.OU(lambda, P, P_1, Sigma, threshold0 = 0)
}
\arguments{
\item{lambda}{vector of the eignevalues of the matrix A of a MV OU process.}

\item{P}{matrix of the eigenvectors of the matrix A}

\item{P_1}{inverse eigenvectors matrix}

\item{Sigma}{the matrix Sigma of a MV OU process.}

\item{Sigmae}{optional additional error (environmental) variance covariance
matrix or single number (if the error has the same variance for all dimensions).
This matrix is added to the MV-OU variance covariance matrix.}
}
\value{
a function of one numerical argument (time), which calculates the
expected variance covariance matrix of a MV-OU process after time, given
the specified arguments.
}
\description{
Generate a multivariate (MV) OU variance-covariance function
}

