% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{V.TwoSpeedOU}
\alias{V.TwoSpeedOU}
\title{Generate a multivariate (MV) TwoSpeedOU variance-covariance function}
\usage{
V.TwoSpeedOU(lambda, P, P_1, Sigma,
  threshold.Lambda_ij = getOption("PCMBase.Threshold.Lambda_ij", 1e-08))
}
\arguments{
\item{lambda}{vector of the eignevalues of the matrix H2 of a MV TwoSpeedOU process.}

\item{P}{matrix of the eigenvectors of the matrix H}

\item{P_1}{inverse eigenvectors matrix}

\item{Sigma}{the matrix Sigma of a MV TwoSpeedOU process.}

\item{threshold.Lambda_ij}{a 0-threshold for abs(Lambda_i + Lambda_j), where Lambda_i
and Lambda_j are eigenvalues of the parameter matrix H2. This
threshold-value is used as a condition to
take the limit time of the expression `(1-exp(-Lambda_ij*time))/Lambda_ij` as
`(Lambda_i+Lambda_j) --> 0`. You can control this value by the global option
"PCMBase.Threshold.Lambda_ij". The default value (1e-8) is suitable for branch
 lengths bigger than 1e-6. For smaller branch lengths, you are may want to
 increase the threshold value using, e.g.
`options(PCMBase.Threshold.Lambda_ij=1e-6)`.}
}
\value{
a function of one numerical argument (time), which calculates the
expected variance covariance matrix of a MV-TwoSpeedOU process after time, given
the specified arguments.
}
\description{
Generate a multivariate (MV) TwoSpeedOU variance-covariance function
}
